Convert to FLOAT32 instead of INT32

+++ qupath-core/src/test/java/qupath/lib/images/servers/TestZProjectionImageServer.java
@@ -317,7 +317,7 @@ public class TestZProjectionImageServer {
 
         @Override
         protected PixelType getPixelTypeWithSum() {
-            return PixelType.INT32;
+            return PixelType.FLOAT32;
         }
 
         @Override
